Added plotslice
